Refactored Url Actor into other classes with prefix multi.
Multi makes it possible to add many url actors to a web page.
This web page creates a server side web actor that has a websocket connection to the web page.
The web actor instantiates and subscribes to many persistent actors in the actor system.

This is a commit before cleanup.